ANOMALY DETECTION,,,.R
The objective of this project is to detect anomalies within a given dataset and assess their impact on the analysis performance. The main focus is on developing a predictive model that will enable accurate sales forecasting. By identifying and addressing anomalies within the dataset, we aim to enhance the overall accuracy and reliability of the sales prediction model. This project seeks to highlight the significance of anomaly detection in improving the quality and effectiveness of sales analysis.